% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_assessments}
\alias{data_assessments}
\title{Assessment data}
\format{
A data frame with 59,260 rows and 7 variables:
\describe{
\item{site}{Assessment site}
\item{event}{Event}
\item{visit_type}{Visit type}
\item{visit_start}{Visit start date}
\item{visit_end}{Visit end date}
\item{days_from_midpoint}{Deviation in days from midpoint of window}
\item{completion_status}{Completion status}
}
}
\usage{
data_assessments
}
\description{
A dataset containing fake assessment data.
}
\keyword{datasets}
