% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_recruitments}
\alias{data_recruitments}
\title{Recruitment data}
\format{
A data frame with 7500 rows and 15 variables:
\describe{
\item{site}{Recruitment site}
\item{date}{Recruitment date}
\item{substance}{Participant is substance user (yes/no)}
\item{race}{Participant's race}
\item{ethnicity}{Participant's ethnicity}
\item{education}{Participant's education level}
\item{poverty}{Participant's poverty level}
\item{ruca}{Participant's RUCA classification}
\item{urban}{Participant's urbanicity classification}
\item{adi}{Participant's adi classification}
\item{opioids}{Participant is opioid user (yes/no)}
\item{marijuana}{Participant is marijuana user (yes/no)}
\item{alcohol}{Participant is alcohol user (yes/no)}
\item{tobacco}{Participant is tobacco user (yes/no)}
\item{stimulants}{Participant is stimulants user (yes/no)}
}
}
\usage{
data_recruitments
}
\description{
A dataset containing fake recruitment data.
}
\keyword{datasets}
