% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii-read.R
\name{pii_read}
\alias{pii_read}
\title{Read PII data and return a dataframe of given fields for a table.}
\usage{
pii_read(
  url = "https://redcap-pii.ucsd.edu/pii/",
  token,
  table,
  fields,
  filter_field = NULL,
  filter_value = NULL,
  batch_size = 100
)
}
\arguments{
\item{url}{The REDCap PII URL.}

\item{token}{Confidential token.}

\item{table}{Name of the PII table.}

\item{fields}{Name of the PII fields.}

\item{filter_field}{Field from 'fields' to filter with. If NULL, return only top ten records in table and specified fields.}

\item{filter_value}{Value of the 'filter_field' to filter with. If NULL, return only top ten records in table and specified fields.}

\item{batch_size}{The maximum number of subject records a single batch should contain.}
}
\value{
tbl. Data frame of PII data.
}
\description{
In use in abcd-ra tools
}
\examples{
\donttest{
pii_read(
   token = "XXX",
   table = "participants_events_log",
   fields = c("pGUID", "eventType", "eventNotes"),
   filter_field = "pguid",
   filter_value = c("YYY", "ZZZ"),
   batch_size = 100
   )

pii_read(
   token = "XXX",
   table = "participants_contacts",
   fields = c("pguid", "contact_id", "relationship", "eventname", "serverdate", "id")
   )
}

}
