% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-categ-plot.R
\name{plot_comp_categ}
\alias{plot_comp_categ}
\title{Compare categorical variables between groups (plot)}
\usage{
plot_comp_categ(
  data_1,
  data_2,
  label_1,
  label_2,
  vars,
  data_dict = NULL,
  type = c("dumbbell", "deviation", "side")
)
}
\arguments{
\item{data_1}{tbl. Data frame with data from the first group.}

\item{data_2}{tbl. Data frame with data from the second group.}

\item{label_1}{character. Label/name for the first group.}

\item{label_2}{character. Label/name for the second group.}

\item{vars}{character (vector). One or several variables/columns to compare
between the groups. The columns are expected to be factor variables. If
they aren't, they are converted to factor.}

\item{data_dict}{tbl. Data frame with columns \code{var} and \code{label} to assign
more descriptive names to the variables in the plot (Optional. If not
provided, the column name will be used).}

\item{type}{character. Type of visualization to create (One of: "dumbbell"
(Default), "deviation", or "side")}
}
\value{
plotly object. The resulting visualization.
}
\description{
High-level function that compares the frequencies of the levels of a
selection of categorical variables between two data frames (i.e., data
from two groups). Allows to choose between three different visualizations:
dumbbell plot; deviation lollipop plot (deviation from 0); side-by-side
lollipop plot. Visualizations are created using \code{ggplot2} and made
interactive using \code{plotly}.
}
\examples{
\dontrun{
# prepare dataset to compare
data_main <- data_recruitments
data_comp <- data_recruitments |>
  slice_sample(
    n = 1000
  )

# compare
plot_comp_categ(
  data_1  = data_main,
  data_2  = data_comp,
  label_1 = "full dataset",
  label_2 = "sampled dataset",

  # try out different combinations (and orders) of variables
  # vars   = c("race", "urban", "ruca", "education"),
  vars    = c("ethnicity", "education", "substance"),

  # try out different plot types
  type    = "dumbbell"
  # type   = "deviation"
  # type   = "side"
)

# use data dictionary to label variables
data_dict <- tribble(
  ~var,         ~label,
  "race",       "Race",
  "ethnicity",  "Ethnicity",
  "substance",  "Substance User",
  "education",  "Highest Education",
  "poverty",    "Poverty Level",
  "ruca",       "RUCA",
  "urban",      "Urbanicity"
)

plot_comp_categ(
  data_1    = data_recruitments,
  data_2    = data_recruitments |>
    slice_sample(
      n = 1000
    ),
  label_1   = "Group 1",
  label_2   = "Group 2",
  data_dict = data_dict,

  # try out different combinations (and orders) of variables
  vars      = c("race", "urban", "ruca", "education"),
  # vars     = c("ethnicity", "education", "substance"),

  # try out different plot types
  # type     = "dumbbell"
  type      = "deviation"
  # type     = "side"
)
}

}
