% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-categ-data.R
\name{prep_data_comp_categ}
\alias{prep_data_comp_categ}
\title{Compare categorical variables between groups (data)}
\usage{
prep_data_comp_categ(
  data_1,
  data_2,
  label_1,
  label_2,
  var,
  type = c("dumbbell", "deviation", "side")
)
}
\arguments{
\item{data_1}{tbl. Data frame with data from the first group.}

\item{data_2}{tbl. Data frame with data from the second group.}

\item{label_1}{character. Label/name for the first group.}

\item{label_2}{character. Label/name for the second group.}

\item{var}{character. Variable/column to compare between the groups. The
column is expected to be a factor variable. If it isn't, it is converted to
factor.}

\item{type}{character. Type of visualization for which the data is being
created (One of: "dumbbell" (Default), "deviation", or "side").}
}
\value{
tbl. The data frame needed to create the selected visualization.
}
\description{
High-level function that prepares the data needed to compare the frequencies
of the levels of a categorical variables between two data frames (i.e.,
data from two groups). Is used internally by \code{DSMviz::plot_comp_categ()}.
Allows to choose to prepare data for three different visualizations:
dumbbell plot; deviation lollipop plot (deviation from 0); side-by-side
lollipop plot.
}
\examples{
# prepare dataset to compare
data_main <- data_recruitments
data_comp <- data_recruitments |>
  dplyr::slice_sample(
    n = 1000
  )

# create data frame for dumbbell plot of `ethnicity` variable
prep_data_comp_categ(
  data_1  = data_main,
  data_2  = data_comp,
  label_1 = "full dataset",
  label_2 = "sampled dataset",
  var    = "ethnicity",
  type    = "dumbbell"
)

# create data frame for deviation plot of `urban` variable
prep_data_comp_categ(
  data_1  = data_main,
  data_2  = data_comp,
  label_1 = "group 1",
  label_2 = "group 2",
  var    = "urban",
  type    = "deviation"
)

}
