% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-categ-data.R
\name{round_any}
\alias{round_any}
\title{Round any}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric (vector). (Vector of) number(s) to round.}

\item{accuracy}{numeric. The accuracy to round to (e.g., specifying 25 will
round to the next increment of 25)}

\item{f}{function. Rounding function to use (Default: round)}
}
\value{
numeric (vector). (Vector of) rounded number(s).
}
\description{
Helper function from the \code{{plyr}} package to ease using rounding functions
(round, floor, ceiling) with specified accuracy.
}
